Fixxed the plotting issue
The issue was that in the initializaiton of the states, i originally
added the properties with the time in a 1x2 array.

then for the tolerancing, i figured itd be best to just delete the time
column to save memory and make it easier to be understood in syntax.

so I went back into the plotting function to change the values in kind.