Finished.
What needs to be done:

STATECLEANER: this needs to be changed to something that includes other
states and also deletes states that to not last longer than 30 seconds.

Figure out how to do multiple columns in the pdf generator.

Figure out how to do page breaks instead of line breaks